♻️ (Transparent Button) - Move transparent button in a separate component
